pyproject.toml
[tool.poetry]
name = "proj"
version = "0.1.0"
description = ""
authors = ["NoFate35 <vsokol35@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
